+++ kOS/Kos Editor/Application/Window.cpp
@@ -131,8 +131,8 @@ namespace Application {
     {
         auto* app = static_cast<Application::AppWindow*>(glfwGetWindowUserPointer(window));
         if (!app) return;
-        app->windowHeight = height;
-        app->windowWidth = width;
+        app->windowHeight = static_cast<float>(height);
+        app->windowWidth = static_cast<float>(width);
         glViewport(0, 0, width, height);
     }
